Issue#446 Scoring based on min-value algo (#448)
Reviewed-on: https://gitea.local.coddicted.com/coddicted/buzzma/pulls/448

+++ backend/src/main/java/com/coddicted/buzzma/claim/dto/ClaimRequestDto.java
@@ -35,5 +35,5 @@ public class ClaimRequestDto {
 
   @NotNull private Map<String, ScoredValue> extractedDetails;
 
-  private Double overallScore;
+  private Integer overallScore;
 }
